(TK-476) Plumb rbac subject extracting function through
This commit plumbs through from the service level to the business logic
the function used to convert an rbac token in a request to an rbac subject.
It also adds a couple of additional function arities that are widely
used in tests.

Fix minor bugs

Validate rbac permission string format

Fixing tests again