(TK-476) Prevent RBAC permissions being used to deny access
The RBAC model assumes all permissions are denied by default, and adding
roles to users allows them to do things. In general, any criterion in
tk-auth can be used either to allow or deny access to an endpoint, but
this doesn't make sense for RBAC permissions. So this commit causes an
exception to be raised if the auth.conf file contains an RBAC permission
being used to deny access.

(TK-476) Add slingshot errors for more clarity

Fix test so it runs